requirements.txt
streamlit==1.28.1
biopython==1.81
requests==2.31.0
pandas==2.1.3 